fix: Handle function details for inlining
When this code looked up function detail, it only took a look at the
first line of a location. A location with multiple lines happens when
functions are inlined, the last entry represents the caller into which the
preceding entries were inlined into.

This corrects this problem and allows to find function detail in this
case as well.